adjust example data docs
instead of maintaining two separate copies of the same source, keep the example data source in an .rmd and include an unevaluated version in the documentation.

this commit also exports the penguins_test and penguins_train data.

also, cut down on gh actions builds again—the macos devel build runs for much longer than the rest.